+++ src/container/App.js
@@ -9,7 +9,7 @@ import {
 import PrivateRoute from './privateRoute';
 
 import Navbar from '../components/common/Layout/Navbar';
-import Footer from '../components/common/Layout/Footer';
+import Footer from '../components/common/Layout//footer/Footer';
 import Sidebar from '../components/common/SideBar';
 
 import LoginPage from '../components/auth/loginPage';
